+++ src/manipulation.js
@@ -1,19 +1,16 @@
 import jQuery from "./core.js";
 import isAttached from "./core/isAttached.js";
-import flat from "./var/flat.js";
 import isIE from "./var/isIE.js";
 import push from "./var/push.js";
 import access from "./core/access.js";
 import rtagName from "./manipulation/var/rtagName.js";
-import rscriptType from "./manipulation/var/rscriptType.js";
 import wrapMap from "./manipulation/wrapMap.js";
 import getAll from "./manipulation/getAll.js";
+import domManip from "./manipulation/domManip.js";
 import setGlobalEval from "./manipulation/setGlobalEval.js";
-import buildFragment from "./manipulation/buildFragment.js";
 import dataPriv from "./data/var/dataPriv.js";
 import dataUser from "./data/var/dataUser.js";
 import acceptData from "./data/var/acceptData.js";
-import DOMEval from "./core/DOMEval.js";
 import nodeName from "./core/nodeName.js";
 
 import "./core/init.js";
@@ -33,21 +30,6 @@ function manipulationTarget( elem, content ) {
 		nodeName( content.nodeType !== 11 ? content : content.firstChild, "tr" ) ) {
 
 		return jQuery( elem ).children( "tbody" )[ 0 ] || elem;
-	}
-
-	return elem;
-}
-
-// Replace/restore the type attribute of script elements for safe DOM manipulation
-function disableScript( elem ) {
-	elem.type = ( elem.getAttribute( "type" ) !== null ) + "/" + elem.type;
-	return elem;
-}
-function restoreScript( elem ) {
-	if ( ( elem.type || "" ).slice( 0, 5 ) === "true/" ) {
-		elem.type = elem.type.slice( 5 );
-	} else {
-		elem.removeAttribute( "type" );
 	}
 
 	return elem;
@@ -83,91 +65,6 @@ function cloneCopyEvent( src, dest ) {
 
 		dataUser.set( dest, udataCur );
 	}
-}
-
-function domManip( collection, args, callback, ignored ) {
-
-	// Flatten any nested arrays
-	args = flat( args );
-
-	var fragment, first, scripts, hasScripts, node, doc,
-		i = 0,
-		l = collection.length,
-		iNoClone = l - 1,
-		value = args[ 0 ],
-		valueIsFunction = typeof value === "function";
-
-	if ( valueIsFunction ) {
-		return collection.each( function( index ) {
-			var self = collection.eq( index );
-			args[ 0 ] = value.call( this, index, self.html() );
-			domManip( self, args, callback, ignored );
-		} );
-	}
-
-	if ( l ) {
-		fragment = buildFragment( args, collection[ 0 ].ownerDocument, false, collection, ignored );
-		first = fragment.firstChild;
-
-		if ( fragment.childNodes.length === 1 ) {
-			fragment = first;
-		}
-
-		// Require either new content or an interest in ignored elements to invoke the callback
-		if ( first || ignored ) {
-			scripts = jQuery.map( getAll( fragment, "script" ), disableScript );
-			hasScripts = scripts.length;
-
-			// Use the original fragment for the last item
-			// instead of the first because it can end up
-			// being emptied incorrectly in certain situations (trac-8070).
-			for ( ; i < l; i++ ) {
-				node = fragment;
-
-				if ( i !== iNoClone ) {
-					node = jQuery.clone( node, true, true );
-
-					// Keep references to cloned scripts for later restoration
-					if ( hasScripts ) {
-						jQuery.merge( scripts, getAll( node, "script" ) );
-					}
-				}
-
-				callback.call( collection[ i ], node, i );
-			}
-
-			if ( hasScripts ) {
-				doc = scripts[ scripts.length - 1 ].ownerDocument;
-
-				// Reenable scripts
-				jQuery.map( scripts, restoreScript );
-
-				// Evaluate executable scripts on first document insertion
-				for ( i = 0; i < hasScripts; i++ ) {
-					node = scripts[ i ];
-					if ( rscriptType.test( node.type || "" ) &&
-						!dataPriv.access( node, "globalEval" ) &&
-						jQuery.contains( doc, node ) ) {
-
-						if ( node.src && ( node.type || "" ).toLowerCase()  !== "module" ) {
-
-							// Optional AJAX dependency, but won't run scripts if not present
-							if ( jQuery._evalUrl && !node.noModule ) {
-								jQuery._evalUrl( node.src, {
-									nonce: node.nonce,
-									crossOrigin: node.crossOrigin
-								}, doc );
-							}
-						} else {
-							DOMEval( node.textContent, node, doc );
-						}
-					}
-				}
-			}
-		}
-	}
-
-	return collection;
 }
 
 function remove( elem, selector, keepData ) {
